+++ WHATSNEW.rst
@@ -62,10 +62,10 @@ Features
   ``number=None`` instead of ``number=False`` (:commit:`f7308cbe`).
 * Use custom locator `proplot.ticker.DiscreteLocator` for major/minor discrete colorbar
   ticks to auto-select subset of levels depending on axis length (:commit:`b94a9b1e`).
+* Refresh major/minor discrete colorbar ticks when the associated axis is drawn to
+  update the ticked levels (:commit:`92bb937e`, :commit:`302c239e`).
 * Register `proplot.ticker.DiscreteLocator` as ``'discrete'`` and add keywords `index`
   and `discrete` to the constructor `~proplot.constructor.Locator` (:commit:`b94a9b1e`).
-* Auto update major/minor discrete colorbar ticks whenever associated axes is drawn
-  to appropriately filter locations (:commit:`92bb937e`, :commit:`302c239e`).
 * Auto disable minor colorbar and axis ticks when major ticks have non-numeric
   labels set by `~matplotlib.ticker.FixedFormatter` (:commit:`c747ae44`).
 * Permit passing `vmin` and `vmax` to `proplot.axes.Axes.colorbar`, as quick